Refactor: Game class split into multiple partial class files.
